third_party/WebKit/LayoutTests/editing/deleting/delete_cell_contents.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => {
    assert_selection(
        '<div contenteditable><table><tr><td>^foo <a href="abc">bar|</a></td><td>baz</td></tr></table></div>',
        'delete',
        '<div contenteditable><table><tbody><tr><td>|<br></td><td>baz</td></tr></tbody></table></div>');
});
</script>
